#22 Added initial structure for Jasmine tests.
This is just a couple of tests for the mediator.js to look at test
project structure, get the libs in, etc. Require.js adds some
complexity to bootstrapping the tests, and I am still looking at better
ways to attach the files under test (right now it is in SpecRunner.html
attaching instances to the window)

This is also more complex as the example and the library files are
fairly intimate, but I'd like to test them as separate efforts.