alendar bugfix
We stop listening to the 'add' event on the collection since it's called
before the model has the time to get an id. Without an id we cannot edit
an event. We now trigger a custom event in the 'create' success callback
in order to have the new id.

The collection didn't update the _byId index, so we do it by hand when a
new model is added.