Update src/config.js
Aura doesn't work in Opera and sometimes falls in IE9.
Opera loads files in a weird way, and when 'core' is required, it loads core.js
instead of mediator.js. This happens due to the fact that RequireJS 
doesn't have anything in require.config.paths. So require config can be global,
thus it works in Opera. And config should be loaded before require.js.